Renovate ignore npm volta
Volta is only local dev dependency, ment mainly if you switch to a older branch with an older node version it uses this version.
If we update this, it will cause a new node install for each new node version, so I would keep this manuall for now.
On CI the node version from the docker image is in charge